requirements.txt
fastapi[standard]==0.116.2
tortoise-orm[asyncpg]==0.21.6
aerich==0.7.2
python-dotenv==1.0.0